add rudimentary https get/post client, fix several errors in the json handling code

+++ src/main/java/com/feed_the_beast/javacurselib/service/logins/login/LoginSession.java
@@ -12,7 +12,7 @@ public class LoginSession {
     public boolean EffectivePremiumStatus;
     public boolean ActualPremiumStatus;
     public int SubscriptionToken;
-    public int Expires;
-    public int RenewAfter;
+    public long Expires;
+    public long RenewAfter;
     public boolean IsTemporaryAccount;
 }
